+++ lib/core_ext/string.rb
